move decl to new file
