+++ src/main/scala/ir/Expr.scala
@@ -355,38 +355,6 @@ case class UninterpretedFunction(name: String, params: Seq[Expr], returnType: IR
   override def toString = s"$name(${params.mkString(", ")})"
 }
 
-sealed trait Decl {
-  def toBoogie: BDeclaration
-}
-case class AxiomDecl(e: Expr) extends Decl {
-  def toBoogie = BAxiom(e.toBoogie, List())
-}
-case class FunctionDecl(
-  name: String,
-  params: List[LocalVar],
-  returnType: IRType,
-  definition: Option[Expr],
-  attribs: List[(String, Option[String])] = List(),
-  inlineDef: Boolean = false
-) extends Decl {
-  def toBoogie =
-
-    val bparams = params.map(p => BParam(p.name, p.getType.toBoogie))
-
-    val body = definition.map(_.toBoogie)
-
-    BFunction(name, bparams, BParam(returnType.toBoogie), body, attribs.map((n, v) => BAttribute(n, v)))
-  def makeCall(actualParams: List[Expr] = List()) = {
-    require(params.map(_.getType) == actualParams.map(_.getType))
-    if (!inlineDef || definition.isEmpty) then {
-      UninterpretedFunction(name, actualParams, returnType, false)
-    } else {
-      val l = LambdaExpr(params, definition.get)
-      ir.eval.evalLambdaApply(l, UninterpretedFunction(name, params, returnType, false))
-    }
-  }
-}
-
 /** Something that has a global scope from the perspective of the IR and Boogie.
   *
   * Not the same as global in the sense of shared memory between threads.
